# src/controller/task.controller.ts
import { Request as ExRequest } from "express";
import {
  Body,
  Delete,
  Get,
  Middlewares,
  OperationId,
  Patch,
  Path,
  Post,
  Put,
  Query,
  Request,
  Response,
  Route,
  Security,
  SuccessResponse,
  Tags,
} from "tsoa";
import BaseController from "./contract/baseController.contract";
import TaskService from "../service/task/task.service";
import { TaskRequest } from "../dto/task/TaskRequest.dto";
import TaskDTO from "../dto/task/task.dto";
import { ValidationErrorResponse } from "../dto/Error/ValidationErrorResponse.dto";
import { UnauthorizedErrorResponse } from "../dto/Error/UnauthorizedErrorResponse.dto";
import { ForbiddenErrorResponse } from "../dto/Error/ForbiddenErrorResponse.dto";
import { validate_schemas } from "../middlewares/validationMiddleware";
import {
  deleteTaskValidationSchema,
  getTaskValidationSchema,
  TaskValidationSchema,
  updateTaskValidationSchema,
} from "./validation/task.validation";
import { authorize } from "../middlewares/authorization";
import { NotFoundError } from "../service/contract/errors/errors";
import { NotFoundErrorResponse } from "../dto/Error/NotFoundErrorResponse.dto";
import { PatchTaskRequest } from "../dto/task/PatchTaskRequest.dto";

// Controller
@Security("jwt")
@Response<ValidationErrorResponse>(422, "Validation failed")
@Response<UnauthorizedErrorResponse>(401, "Unauthorized")
@Response<ForbiddenErrorResponse>(403, "Forbidden")
@Response<NotFoundErrorResponse>(404, "Not Found Error")
@Route("/api/tasks")
@Tags("Tasks")
export class TaskController extends BaseController {
  /**
   * Create New Task
   * @param req  Express Request including user info
   * @param body Task creation data including name, status, priority, etc.
   * @returns The created task response payload
   */
  @SuccessResponse("200", "Task Created Successfully")
  @Middlewares([
    validate_schemas(TaskValidationSchema, "body"),
    authorize("create_task"),
  ])
  @Post()
  async create(@Request() req: ExRequest, @Body() body: TaskRequest) {
    const taskData = {
      ...body,
      startDate: body.startDate ? new Date(body.startDate) : undefined,
      endDate: body.endDate ? new Date(body.endDate) : undefined,
    };
    const currentUserUid = req.user.uid;
    console.log("Current user from controller: ", req.user.uid);

    const task = await TaskService.create(currentUserUid, taskData);

    return super.postOk({
      message: "Task Created Successfully",
      data: TaskDTO.single(task),
    });
  }

  @SuccessResponse("200", "Task Retrieved Successfully")
  @Get()
  @Middlewares(authorize("read_task"))
  async getAll(
    @Request() req: ExRequest,
    @Query() page: number = 1,
    @Query() limit: number = 10
  ) {
    const { tasks, total } = await TaskService.getAllPaginated(page, limit);

    return Object.assign(
      super.getOk({
        message: total ? "Tasks fetched successfully" : "No Task found",
        data: tasks,
      }),
      { total }
    );
  }

  @SuccessResponse("200", "Task Retrieved Successfully")
  @Middlewares([
    validate_schemas(getTaskValidationSchema, "params"),
    authorize("read_task"),
  ])
  @Get("{taskUid}")
  public async getByUid(
    // @Request() req: ExRequest,
    @Path() taskUid: string
  ) {
    const task = await TaskService.getByUid(taskUid);

    return super.getOk({
      message: "Task Retrieved Successfully",
      data: this.serializeTask(task),
    });
  }

  @SuccessResponse("200", "Task Updated Successfully")
  @Middlewares([
    validate_schemas(updateTaskValidationSchema, "body"),
    validate_schemas(getTaskValidationSchema, "params"),
    authorize("update_task"),
  ])
  @Put("{taskUid}")
  public async update(
    @Request() req: ExRequest,
    // req: Request,
    @Path() taskUid: string,
    @Body() body: TaskRequest
  ) {
    const updateData = {
      ...body,
      startDate: body.startDate ? new Date(body.startDate) : undefined,
      endDate: body.endDate ? new Date(body.endDate) : undefined,
    };

    const updatedTask = await TaskService.updateTask(taskUid, updateData);

    return super.putOk({
      message: "Task Updated Successfully",
      data: this.serializeTask(updatedTask),
    });
  }

  private serializeTask(task: any) {
    return {
      ...task,
      deadline: task.deadline?.toISOString() ?? null,
      createdAt: task.createdAt.toISOString(),
      updatedAt: task.updatedAt.toISOString(),
    };
  }

  @SuccessResponse("200", "Task Updated Successfully")
  @Middlewares([
    validate_schemas(updateTaskValidationSchema, "body"),
    validate_schemas(getTaskValidationSchema, "params"),
    authorize("update_task_status"),
  ])
  @Patch("{taskUid}")
  public async patch(
    @Request() req: ExRequest,
    // req: Request,
    @Path() taskUid: string,
    @Body() body: PatchTaskRequest
  ) {
    const user = req.user;

    const patchData = {
      ...body,
      startDate: body.startDate ? new Date(body.startDate) : undefined,
      endDate: body.endDate ? new Date(body.endDate) : undefined,
    };

    const updatedTask = await TaskService.patch(user.uid, taskUid, patchData);

    return super.putOk({
      message: "Task Updated Successfully",
      data: this.serializeTask(updatedTask),
    });
  }

  @SuccessResponse("200", "Task Deleted Successfully")
  // forbidden 403
  @Middlewares([
    validate_schemas(deleteTaskValidationSchema, "params"),
    authorize("delete_task"),
  ])
  @Delete("{taskUid}")
  public async delete(@Request() request: ExRequest, @Path() taskUid: string) {
    // const userId = request.user.id; // Verify user is authenticated
    await TaskService.deleteTask(taskUid, request.user);
    return super.deleteOk({
      message: "Task Deleted Successfully",
      data: {},
    });
  }
}